+++ script_templates/Node/ExtensibleCommand.cs
@@ -4,28 +4,22 @@ using _BINDINGS_NAMESPACE_;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
+using static YAT.Scenes.BaseTerminal.BaseTerminal;
 
-namespace YAT.Commands
+[Command("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_")]
+public partial class _CLASS_ : Extensible, ICommand
 {
-	[Command("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_")]
-	public partial class _CLASS_ : Extensible, ICommand
+	public CommandResult Execute(CommandArguments args)
 	{
-		public YAT Yat { get; set; }
+		var option = args.Arguments[1];
 
-		public _CLASS_(YAT Yat) => this.Yat = Yat;
-
-		public CommandResult Execute(YAT yat, params string[] args)
+		if (Extensions.ContainsKey(option))
 		{
-			var option = args[1];
-
-			if (Extensions.ContainsKey(option))
-			{
-				var extension = Extensions[option];
-				return extension.Execute(yat, this, args[1..]);
-			}
-
-			yat.Terminal.Println("Options not found.", Terminal.PrintType.Error);
-			return CommandResult.Failure;
+			var extension = Extensions[option];
+			return extension.Execute(args with { Arguments = args.Arguments[1..] });
 		}
+
+		args.Terminal.Println("Options not found.", PrintType.Error);
+		return CommandResult.Failure;
 	}
 }