+++ script_templates/Node/Extension.cs
@@ -4,13 +4,14 @@ using _BINDINGS_NAMESPACE_;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
+using static YAT.Scenes.BaseTerminal.BaseTerminal;
 
 [Extension("_CLASS_", "Lorem ipsum dolor sit amet.", "[b]Usage[/b]: _CLASS_")]
 public partial class _CLASS_ : IExtension
 {
-	public CommandResult Execute(YAT.YAT yat, ICommand command, params string[] args)
+	public CommandResult Execute(CommandArguments args)
 	{
-		yat.Terminal.Print("_CLASS_ is not yet implemented!", Terminal.PrintType.Warning);
+		args.Terminal.Print("_CLASS_ is not yet implemented!", PrintType.Error);
 
 		return CommandResult.Failure;
 	}